Déplace SubCategoryFactory et LicenseFactory dans le module utils
Rendre à César ce qui lui appartient
