+++ zds/notification/tests/tests_tricky.py
@@ -20,13 +20,12 @@ from zds.notification.models import (
 from zds.tutorialv2 import signals
 from zds.tutorialv2.factories import (
     PublishableContentFactory,
-    LicenceFactory,
-    SubCategoryFactory,
     PublishedContentFactory,
     ContentReactionFactory,
 )
 from zds.tutorialv2.publication_utils import publish_content, notify_update
 from zds.tutorialv2.tests import TutorialTestMixin, override_for_contents
+from zds.utils.factories import SubCategoryFactory, LicenceFactory
 from zds.utils.mps import send_mp, send_message_mp
 from zds.utils.header_notifications import get_header_notifications
