+++ zds/tutorialv2/tests/tests_front.py
@@ -15,15 +15,13 @@ from selenium.webdriver.support.ui import WebDriverWait
 
 from zds.member.factories import StaffProfileFactory, ProfileFactory
 from zds.tutorialv2.factories import (
-    LicenceFactory,
-    SubCategoryFactory,
     PublishableContentFactory,
     ContainerFactory,
     ExtractFactory,
 )
 from zds.tutorialv2.models.database import PublishedContent, PublishableContent
 from zds.tutorialv2.tests import TutorialTestMixin, TutorialFrontMixin
-from zds.utils.factories import CategoryFactory
+from zds.utils.factories import CategoryFactory, SubCategoryFactory, LicenceFactory
 
 overridden_zds_app = deepcopy(settings.ZDS_APP)
 overridden_zds_app["content"]["repo_private_path"] = settings.BASE_DIR / "contents-private-test"
