+++ zds/tutorialv2/tests/tests_models.py
@@ -14,14 +14,13 @@ from zds.tutorialv2.factories import (
     PublishableContentFactory,
     ContainerFactory,
     ExtractFactory,
-    LicenceFactory,
     PublishedContentFactory,
-    SubCategoryFactory,
 )
 from zds.gallery.factories import UserGalleryFactory
 from zds.tutorialv2.models.database import PublishableContent, PublishedContent
 from zds.tutorialv2.publication_utils import publish_content
 from zds.tutorialv2.tests import TutorialTestMixin, override_for_contents
+from zds.utils.factories import SubCategoryFactory, LicenceFactory
 from zds.utils.models import Tag
 from django.template.defaultfilters import date
