+++ zds/tutorialv2/tests/tests_opinion_views.py
@@ -8,12 +8,11 @@ from zds.member.factories import ProfileFactory, StaffProfileFactory
 from zds.tutorialv2.factories import (
     PublishableContentFactory,
     ExtractFactory,
-    LicenceFactory,
     PublishedContentFactory,
-    SubCategoryFactory,
 )
 from zds.tutorialv2.models.database import PublishableContent, PublishedContent, PickListOperation
 from zds.tutorialv2.tests import TutorialTestMixin, override_for_contents
+from zds.utils.factories import SubCategoryFactory, LicenceFactory
 from zds.utils.models import Alert
