+++ zds/tutorialv2/tests/tests_views/tests_editcontentlicense.py
@@ -7,7 +7,8 @@ from zds.tutorialv2.views.contents import EditContentLicense
 from zds.tutorialv2.forms import EditContentLicenseForm
 from zds.tutorialv2.tests import TutorialTestMixin, override_for_contents
 from zds.member.factories import ProfileFactory, StaffProfileFactory
-from zds.tutorialv2.factories import PublishableContentFactory, LicenceFactory
+from zds.tutorialv2.factories import PublishableContentFactory
+from zds.utils.factories import LicenceFactory
 
 
 @override_for_contents()
